move extensions to .proposed
